% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enciphR.R
\name{enciphR}
\alias{enciphR}
\title{enciphR}
\usage{
enciphR(x, alg, key = F)
}
\arguments{
\item{x}{string to be enciphered}

\item{alg}{algorithm for cipher (e.g. 1 to shift all letter forward 1 \verb{[A->B]}; -2 to shift all letters back 2 places \verb{[A->Y]})}

\item{key}{do you want a key to be included in output?; default=F; If key=T, output will be a list, rather than a vector.}
}
\description{
encipher any text string using an alphabetic shift
}
