% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhyloNames.R
\name{getPhyloNames}
\alias{getPhyloNames}
\title{getPhyloNames}
\usage{
getPhyloNames(speciesNames, nameType, clearCache = F, quiet = T)
}
\arguments{
\item{speciesNames}{a name or vector of common or scientific names of organisms of interest (in quotes)}

\item{nameType}{what type of name are you supplying? Either "sci" or "common"}

\item{clearCache}{delete the cached phylonamescache.rds file saved in tempdir()? default=F}

\item{quiet}{suppress verbose feedback from the taxize package? Default=T}
}
\description{
Find matching scientific or common names for the common or scientific names provided and cache the results for efficient retrieval. This is a convenience wrapper for the \code{\link[taxize]{sci2comm}} and \code{\link[taxize]{comm2sci}} functions.
}
\details{
Depending on what nameType you supply, getPhyloNames will use the sci2comm or comm2sci function to look for the matching taxonomic name. It first searches the NCBI database, and if that doesn't return a result, it will try the Encyclopedia of Life (EOL) database. This function relies on an internal helper function \code{\link{getPhyloNames_noCache}}, though you will generally not need to use it. The advantage of getPhyloNames is that it caches results, since the database APIs can be a little slow, so you will not need to keep looking up the same names over and over again.
}
