% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhyloNames_noCache.R
\name{getPhyloNames_noCache}
\alias{getPhyloNames_noCache}
\title{getPhyloNames_noCache}
\usage{
getPhyloNames_noCache(speciesNames, nameType, quiet = T)
}
\arguments{
\item{speciesNames}{a name or vector of common or scientific names of organisms of interest (in quotes)}

\item{nameType}{what type of name are you supplying? Either "sci" or "common"}

\item{quiet}{suppress verbose feedback from the taxize package? Default=T}
}
\description{
Internal helper function for getPhyloNames; looks for either a scientific or common name matching whichever you supply, based on nameType. Does the work of interfacing with the \code{\link[taxize]{sci2comm}} and \code{\link[taxize]{comm2sci}} functions
}
\details{
You generally won't want or need to use this function, since \code{\link{getPhyloNames}} invokes it and will cache the results to run more efficiently. But this is basically a convenience wrapper that can suppress some of the messaging from taxize and also search both NCBI and EOL databases for taxonomic matches.
}
