% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggciphR.R
\name{ggciphR}
\alias{ggciphR}
\title{ggciphR}
\usage{
ggciphR(ggGraph, alg)
}
\arguments{
\item{ggGraph}{a ggplot2 object}

\item{alg}{algorithm for cipher (e.g. 1 to shift all letter forward 1 \verb{[A->B]}; -2 to shift all letters back 2 places \verb{[A->Y]})}
}
\description{
encipher the axis labels of a ggplot2 object to turn a graph into a data literacy puzzle
}
