% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_labs.R
\name{highlight_labs}
\alias{highlight_labs}
\title{highlight_labs}
\usage{
highlight_labs(which_lab, txt_col = "#005792", font_face = 4, ...)
}
\arguments{
\item{which_lab}{Which label to highlight? Accepts a vector of 1 or more numbers or strings:
\itemize{
\item 1 = "title"
\item 2 = "x-lab"
\item 3 = "y-lab
\item 4 = "x-vals"
\item 5 = "y-vals"
\item 6 = "legend title"
\item 7 = "legend text"
\item 8 = "all"
}}

\item{txt_col}{Color of label text; default= "#005792" (Dark GP Hydrogen Blue)}

\item{font_face}{style of highlighted labels and title fonts; 1=plain, 2= bold, 3=italic, 4=bold+italic; default= 4}
}
\description{
The counterpoint to \code{\link[=blank_labs]{blank_labs()}}, this function makes selected labels bold, italic, and colorful.
}
\details{
This is used to make teacher graphs where it's clear what students will fill in on the blanked graph.
}
\examples{
require(ggplot2)
ggplot(mtcars,aes(x=mpg,y=hp))+geom_point()+
theme_galactic()+ggtitle("Normal graph")

ggplot(mtcars,aes(x=mpg,y=hp))+geom_point()+
theme_galactic()+highlight_labs(c(2,3))
}
\seealso{
Other functions for enigmatizing graphs: 
\code{\link{blank_labs}()}
}
\concept{functions for enigmatizing graphs}
