% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPhylo.R
\name{showPhylo}
\alias{showPhylo}
\title{showPhylo}
\usage{
showPhylo(
  speciesNames,
  nameType,
  dateTree = T,
  labelOffset = 0.45,
  aspectRatio = 1,
  pic = "wiki",
  dotsConnectText = F,
  picSize = 1,
  picSaveDir,
  optPicWidth = 200,
  picBorderWidth = 10,
  picBorderCol = "#363636",
  openDir = F,
  xAxisPad = 0.2,
  xTitlePad = 20,
  numXlabs = 8,
  textScalar = 1,
  xTitleScalar = 1,
  phyloThickness = 1.2,
  phyloCol = "#363636",
  textCol = "#363636",
  plotMar = c(t = 0.02, r = 0.36, b = 0.02, l = 0.02),
  clearCache = F,
  quiet = T
)
}
\arguments{
\item{speciesNames}{a name or vector of common or scientific names of organisms of interest (in quotes)}

\item{nameType}{what type of name are you supplying? Either "sci" or "common"}

\item{dateTree}{try to scale the tree to estimated divergence times? Uses \code{\link[datelife]{datelife_search}} with summary_format="phylo_median." default= T}

\item{labelOffset}{how far from the tree tips do you want to put labels? default=0.3 (in proportion of x-axis units)}

\item{aspectRatio}{doesn't actually work yet; the output phylogeny is always square for the moment}

\item{pic}{what type of species image do you want to plot? options="wiki" (Wikipedia page profile image), "phylopic" (species' silhouette from the PhyloPic repository), "cust" (custom images: must be named .jpg or .png with names matching speciesNames in the picSaveDir folder), or "none"}

\item{dotsConnectText}{do you want a dotted line to go from the text to the labels? default=F}

\item{picSize}{how big to scale images, where 1=100\%; .5=50\%; default=1}

\item{picSaveDir}{location for saving downloaded images; default=fs::path(tempdir(),"showPhylo")}

\item{optPicWidth}{picture width in pixels for optimized versions of images saved if using pic="cust"; default= 200}

\item{picBorderWidth}{for pic="wiki" or "cust," what size border would you like around your image (as a \%); default=10}

\item{picBorderCol}{color of image border; default="#363636" (a dark, near-black color)}

\item{openDir}{for pic="wiki" or "cust," do you want to open the picSaveDir after processing files? default=F}

\item{xAxisPad}{spacing between the phylogeny and the x-axis (scale is 1 is the distance between two sister taxa); default=.2}

\item{xTitlePad}{spacing between x-axis title and x numbers; default=20}

\item{numXlabs}{the number of year markers on the x-axis; default=8}

\item{textScalar}{multiplier of text size for labels and axis numbers; default=1}

\item{xTitleScalar}{multiplier of x-axis title size; default=1}

\item{phyloThickness}{how thick to make the phylogeny lines; default=1.2}

\item{phyloCol}{color of the phylogeny lines; default= "#363636"}

\item{textCol}{color of the axis and tip labels; default= "#363636"}

\item{plotMar}{margins around the plot area in proportional screen width units; note the right margin is much wider to make room for tip labels; default=c(t=.02,r=.1,b=.02,l=.02) for top, right, bottom, left}

\item{clearCache}{delete cached images and taxonomic names? Passed to getPhyloNames, getWikiPics, and also applies to optimized custom images; default=F}

\item{quiet}{suppress verbose feedback from the taxize package? Passed to getPhyloNames and get WikiPic helper functions. Default=T}
}
\description{
A \emph{relatively} simple function for generating presentation-ready phylogenies, given a set of common or scientific names. It works best for a few species, since the goal is to make a phylogeny for pedagogical purposes, so the focus is on details, not quantity.
}
\details{
You have the ability to look up matching common or scientific names for taxa, plot them in an evolutionary tree based on the Open Tree of Life Project, date taxa using datelife, if data are available, plot a 'phylopic' (i.e. a silhouette of the organism) if available, pull down the profile image from the organisms Wikipedia page, or provide a custom image, and label the pictures with the scientific and common names. The result is a customizable ggplot2 object. This is a convenience wrapper that merges functionality from \pkg{ggplot2}, \pkg{ggtree}, \pkg{datelife}, \pkg{rotl},and \pkg{taxize}, among other packages.
}
